v3.0/requirements.txt
Flask
firebase-admin
PyJWT
bcrypt
python-dotenv